scripts/requirements.txt
torch
torchvision
diffusers
transformers
pillow
accelerate